+++ packages/style-utils/CHANGELOG.md
@@ -1,6 +1,38 @@
 # @granite-js/style-utils
 
-## 0.1.23-next.19
+## 0.1.33
+
+## 0.1.32
+
+### Patch Changes
+
+- 7572713: bump version up babel
+
+## 0.1.31
+
+## 0.1.30
+
+## 0.1.29
+
+## 0.1.28
+
+### Patch Changes
+
+- 1d958cc: ci: provenance
+
+## 0.1.27
+
+### Patch Changes
+
+- e32b020: ci: oidc
+
+## 0.1.26
+
+### Patch Changes
+
+- 41f9ab3: ensure polyfill setup during initialization and update package.json config for improve tree shaking
+
+## 0.1.25
 
 ### Patch Changes